Update move
This was causing a bug where a newly promoted piece was being deleted if it put the opposing king in check and the only way to get out of check was to capture it. This is because, when the game looks to see if a player is in checkmate, it checks to see if it can capture the attacking piece. When doing so, it calls this move method to move to the square of the attacking piece, which, by default, deletes that piece. The upshot was that the player in check could move anywhere because, internally, the attacking piece had already been deleted when the game called the checkmate? method.

To fix this, this updates the move method so that it doesn't delete the piece if its caller contains 'checkmate?'